research/online-decoder-client/python/requirements.txt
websocket-client==1.1.0
librosa==0.8.1
numpy==1.21.0

